add skeleton, add toast
