+++ package.json
@@ -20,14 +20,16 @@
     "@tanstack/react-query-devtools": "^5.91.1",
     "@vitejs/plugin-react": "^5.1.2",
     "autoprefixer": "^10.4.23",
+    "clsx": "^2.1.1",
+    "daisyui": "^5.5.14",
     "eslint-import-resolver-typescript": "^4.4.4",
     "eslint-plugin-jsx-a11y": "^6.10.2",
     "eslint-plugin-react-hooks": "^7.0.1",
-    "daisyui": "^5.5.14",
     "jwt-decode": "^4.0.0",
     "react": "^19.2.3",
     "react-dom": "^19.2.3",
     "react-hook-form": "^7.69.0",
+    "react-hot-toast": "^2.6.0",
     "react-intersection-observer": "^10.0.0",
     "react-router-dom": "^7.11.0",
     "tailwindcss": "^4.1.18",
